Add TransferList file
